Added feature to sort todo portlet table and fixed the scrolling on it
Added feature to sort todo portlet table and fixed the scrolling on it. I also made the controller for the todo portlet return with a sorted list by date to be initially displayed. For the sorting I used a merge sort with the ability to change sorting order ascending/descending with the use of a compare function. It is also pretty flexible to add not columns to sort or not and switch positions of columns.